secmet: move common.test.helpers.DummyCDS to secmet where it belongs
