+++ antismash/common/secmet/test/helpers.py
@@ -1,0 +1,24 @@
+# License: GNU Affero General Public License v3 or later
+# A copy of GNU AGPL v3 should have been included in this software package in LICENSE.txt.
+
+""" Simple constructors for complicated features to simplify testing """
+
+# for test files, silence irrelevant and noisy pylint warnings
+# pylint: disable=no-self-use,protected-access,missing-docstring
+
+from ..features.cds_feature import CDSFeature
+from ..locations import FeatureLocation
+
+
+class DummyCDS(CDSFeature):
+    counter = 0
+
+    def __init__(self, start=0, end=7, strand=1, locus_tag=None, translation=None):
+        if not translation:
+            translation = "A"*(abs(start-end))
+        if not locus_tag:
+            locus_tag = "dummy_locus_tag_%d" % DummyCDS.counter
+            DummyCDS.counter += 1
+        super().__init__(FeatureLocation(start, end, strand), translation=translation,
+                         locus_tag=locus_tag)
+        assert self.get_accession() == locus_tag, self.get_accession()
